test: split symlink tests to avoid CI OOM
